web/dashboard-api/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-jose==3.3.0
pyjwt